Fix a major bug where IDs were being written incorrectly.
Due to the Uint64ID BinaryMarshaller containing a bug, all IDs
were written prefixed with 8 extra zeroed out bytes. This bug only
became apparent when it was observed that WakeAt was not working
correctly.

Delay was impacted by this bug, but not Send or Wait.

Along with fixing the bug in Delay, performance should improve
marginally.

Fixes #7
Fixes #8