Add int64Fifo, an on-disk fifo implemented as two stacks.
This will form the basis of the new queueing system; the bold db
will only be used as object storage; the fifo elements will be used
as pointers to the bolt objects.

This would hopefully reduce the amount of latency in the system, as
it would avoid bolt transactions when moving between states like
ready and unacked.