Replace boltdb with modernc.org/sqlite
This commit replaces the storage backend of lasr. Now, sqlite is used.

It also removes support for dead-lettering, putting the responsibility
for that in the hands of the user.

The new storage backend comes with a significant performance penalty.
The library is now around 10 times slower. More work will be needed to
achieve parity with the boltdb-backed queue.

Signed-off-by: Eric Chlebek <eric@sensu.io>